Normalise on a single shared helper
