+++ src/primitives/storyboards.ts
@@ -1,4 +1,4 @@
-import { getMuxStoryboardBaseUrl } from "@mux/ai/lib/mux-image-url";
+import { getMuxStoryboardBaseUrl } from "@mux/ai/lib/mux-url";
 import { signUrl } from "@mux/ai/lib/url-signing";
 import type { WorkflowCredentialsInput } from "@mux/ai/types";
